GH-5: Add settings.json and expand CMS backend ADR
this part of #5

Introduce a new settings.json file for .fleet with properties to enhance backend performance and format code on save. Update the CMS backend ADR document with context about content management needs for a portfolio site and include implementation details for WordPress with pros and cons for using it in the project.