GH-5: Choose TinaCMS as CMS backend and update ADR.
this close #5
TinaCMS was selected because it offers first-class support for persisting content over files instead of using a database and provides API access to the content. The decision outlines both advantages, like easy content editing and GitHub visibility, and drawbacks, such as complex configuration due to auto-generated clients. Additional sections were added for DecapCMS with emphasis on its disadvantages.